moved list and stack to separate container module
