+++ src/common/containers/list.c
@@ -7,7 +7,7 @@
 
 #include "amidilib.h"
 #include "amlog.h"
-#include "list/list.h"
+#include "containers/list.h"
 
 #include "timing/mfp.h"
