+++ src/common/containers/stack.c
