+++ src/tos/midi_rep.c
@@ -12,7 +12,7 @@
 #include "timing/miditim.h"
 #include "timing/mfp.h"
 #include "midi_send.h"
-#include "list/list.h"
+#include "containers/list.h"
 
 #include <stdio.h>
